+++ mipscripts/__main__.py
@@ -4,6 +4,7 @@ import argparse
 import re
 import os
 import os.path
+import pandas
 import pickle
 import sys
 import subprocess
@@ -202,8 +203,36 @@ def seqrun_stats(args):
         # Convert header to snake case
         header_to_snake_case(path=samplesheet, overwrite=True)
 
-        samplesetnum = collections.Counter()
-        print("########### FASTQS / READS #################")
+        # Initialize a counter for the number of sample sets
+        num_sample_sets = collections.Counter()
+
+        # Initialize the grouping dictionary
+        groupings = {}
+
+        # Load the sample sheet header and data into memory for fast retrieval
+        # Also update the number of sample per sample set
+        sampledict = []
+        header = []
+        with open(samplesheet) as samplefile:
+            dictreader = csv.DictReader(samplefile, delimiter="\t")
+            header = dictreader.fieldnames
+            for row in dictreader:
+                sampledict.append(row)
+                num_sample_sets.update({row["sample_set"]: 1})
+
+        # Error if arguments are not correct
+        if not args.maingrp in header:
+            print("\nERROR: --maingrp invalid.")
+            print(f"\u2022 User entered '{args.maingrp}'.")
+            print(f"\u2022 Available options: {header}.")
+            exit()
+        if args.subgrp and not args.subgrp in header:
+            print("\nERROR: --subgrp invaluid.")
+            print(f"\u2022 User entered '{args.maingrp}'.")
+            print(f"\u2022 Available options: {header}.")
+            exit()
+
+        print("####################### FASTQS / READS ########################")
         fastqdir = os.path.dirname(samplesheet) + "/fastq"
         fastqs = os.listdir(fastqdir)
         fqss = [f for f in fastqs if "_R1_" in f]
@@ -220,39 +249,23 @@ def seqrun_stats(args):
             fastqlen[fq] = reads
             bar.title("Counting the number of FASTQ reads")
 
-        # Load the table into memory for fast retrieval
-        sampledict = []
-        header = []
-        with open(samplesheet) as samplefile:
-            dictreader = csv.DictReader(samplefile, delimiter="\t")
-            for row in dictreader:
-                sampledict.append(row)
-            header = dictreader.fieldnames
+        # Save all fastq file names
+        fqfiles = fastqlen.keys()
 
-        # Error if arguments are not correct
-        if not args.maingrp in header:
-            print("\nERROR: --maingrp invalid.")
-            print(f"\u2022 User entered '{args.maingrp}'.")
-            print(f"\u2022 Available options: {header}.")
-            exit()
-        if args.subgrp and not args.subgrp in header:
-            print("\nERROR: --subgrp invaluid.")
-            print(f"\u2022 User entered '{args.maingrp}'.")
-            print(f"\u2022 Available options: {header}.")
-            exit()
-
+        # For each row in our data, check if there are multiple fastqs and, if
+        # not, set the read count
         for row in sampledict:
+            # Define identifier to compare fastqs
             fqname = "{}-{}-{}_".format(
                 row["sample_name"], row["sample_set"], row["replicate"]
             )
-            fqfiles = fastqlen.keys()
-            row["read_count"] = 0
 
             # Use regex to check if the fastq name appears multiple times
             regex = re.compile("^" + fqname)
             fqmatchs = list(filter(regex.match, fqfiles))
 
             if len(fqmatchs) > 1:
+                # Throw an error
                 print(
                     "ERROR: More than one FASTQ for the sampleset rep.",
                     fqmatchs,
@@ -266,18 +279,25 @@ def seqrun_stats(args):
                 )
                 exit()
             elif len(fqmatchs) == 1:
+                # Set read count
                 row["read_count"] = fastqlen[fqmatchs[0]]
 
             # Create groupings
-            groupings = {}
+            # TODO - we don't do anything with the grouping variables
             groupings[row[args.maingrp]] = {}
             if args.subgrp:
                 groupings[row[args.maingrp]][row[args.subgrp]] = 1
 
-        # Create updated samplesheet
+        # If there were undetermined reads let the user know
+        if "Undetermined_S0_R1_001.fastq.gz" in fastqlen:
+            cnt = fastqlen["Undetermined_S0_R1_001.fastq.gz"]
+            print(f"There were {cnt} undetermined reads from demultipelxing.")
+
+        # Create updated sample sheet
         print("Creating an updated samplesheet with read counts.")
         sheet_root = os.path.splitext(samplesheet)[0]
-        with open(f"{sheet_root}_readcnt.tsv", mode="wt") as samplesheetout:
+        read_cnt_path = f"{sheet_root}_readcnt.tsv"
+        with open(read_cnt_path, mode="wt") as samplesheetout:
             header.append("read_count")
             dictwriter = csv.DictWriter(
                 samplesheetout, delimiter="\t", fieldnames=header
@@ -286,36 +306,33 @@ def seqrun_stats(args):
             for row in sampledict:
                 dictwriter.writerow(row)
 
-        print("########### SAMPLES #################")
-        if "Undetermined_S0_R1_001.fastq.gz" in fastqlen:
-            print(
-                "Undetermined reads from demultipelxing",
-                fastqlen["Undetermined_S0_R1_001.fastq.gz"],
-            )
-        print(f"SAMPLE GROUPINGS AND # SAMPLES: {groupings}")
+        print("########################### SAMPLES ###########################")
+        print("PER SAMPLE SET SUMMARY:")
+        # print(f"SAMPLE GROUPINGS AND # SAMPLES: {groupings}")
 
-        # THIS CODE DOES NOT RUN AS THERE IS NO SAMPLESET VARIABLE DEFINED
-        for sampleset in list(samplesetnum.keys()):
-            fqwithreads = len(samplereads.values())
-            total_reads = sum(samplereads.values())
-            mean_reads = int(total_reads / fqwithreads)
-            median_reads = statistics.median(samplereads.values())
-            quantiles = [mean_reads]
-            if fqwithreads > 1:  # quantiles can handle 1 data point
+        # Print some summary stats on reads for each sample set.
+        # Read the read count file in
+        read_cnt_data = pandas.read_csv(read_cnt_path, sep="\t")
+        for key in num_sample_sets:
+            # Print sample set name and number of samples
+            print(f"Sample set {key} ({num_sample_sets[key]:,} samples):")
+
+            # Filter data to sample set and isolate reads
+            filter_data = read_cnt_data[read_cnt_data.sample_set == key]
+            num_samples = filter_data.shape[0]
+            reads = filter_data.read_count
+
+            # Print summary stats
+            print(f"... TOTAL read pairs: {sum(reads):,}")
+            print(f"... MEAN read pairs: {int(statistics.mean(reads)):,}")
+            print(f"... MEDIAN read pairs: {statistics.median(reads):,}")
+
+            # Deciles
+            if num_samples > 1:
                 quantiles = [
-                    round(q)
-                    for q in statistics.quantiles(samplereads.values(), n=10)
+                    f"{round(q):,}" for q in statistics.quantiles(reads)
                 ]
-                quantiles = [f"{q:,}" for q in quantiles]
-            print(f"... TOTAL read pairs {total_reads:,}")
-            print(f"... MEAN read pairs {mean_reads:,}")
-            print(
-                f"... MEDIAN read pairs RS {median_reads:,}",
-                "... DECILES ",
-                quantiles,
-                "median is 50%",
-            )
-            # print (samplereads)
+                print(f"... QUANTILES for read pairs: {quantiles}")
 
 
 #######################################################################3