+++ mipscripts/__main__.py
@@ -5,20 +5,13 @@ import re
 import os
 import os.path
 import pandas
-import pickle
 import sys
-import subprocess
 import textwrap
 import statistics
-import glob
 import copy
 import csv
 import gzip
 import collections
-from collections import defaultdict, Counter, OrderedDict
-import numpy
-import scipy
-import scipy.interpolate
 from alive_progress import alive_it
 
 from .utils import *