+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## mipscripts (development version)
 
+- Ensure standard column names by converting names to snake case
+  ([#3](https://github.com/bailey-lab/mipscripts/issues/3))
+
 ## mipscripts 0.3.2
 
 - Install dependencies when installing package.